Addition of LDACBT_EQMID_ABR with value 0x7F
The value 0x7F is used to specify the LDAC adaptation bit rate in the
current implementation. However, libldac does not have an ID for this
use case. Therefore, there was the possibility in the future that
libldac reuses the value 0x7F for some other LDACBT_EQMID_ purpose.
To avoid this possibility, new entry LDACBT_EQMID_ABR with value 0x7F
added.

Change-Id: I0c44a7906c758f73d554d7166c09c5e890083bf7
Signed-off-by: Chisato Kenmochi <Chisato.Kenmochi@sony.com>